facial animations seem way too intense to do
The way facials work from my guess is by remapping the bones to work as "sliders" essentially. The thing that is remapping them into "sliders" would most likely be the expression dictionary. I did a dumb thing of using one viseme/facial bone at a time to determine what is what. Many are small lip movements and I didn't doc/note much of it as I should've. 840 is the jaw. "Track" 22 seem to be the viseme/facial bones.